refactor: global-error の静的スタイルとテキストを外出しする
ERROR_TEXT / FALLBACK_STYLES を_global-error/fallback-content.tsへ、
ErrorButtonをコンポーネントファイルへ分離し、global-error.tsxは
合成のみに縮小する(243行→172行)。captureClientBoundaryErrorによる
Sentry正規化は変更しない。

Refs #1532
